Refactored Stage
Stage Button & StageUsers in own Component in subfolder of Stage addon
